Fix R64.RANGEINTARRAY random result (#117)
* add support for 64-bit #110

* fix: correct reply handling for 64-bit bitmap commands

Use `uint64_t` and `RedisModule_ReplyWithBigNumber` to handle cases
where bit positions exceed `long long` range, ensuring proper replies
for large values.

* fix: bitmap64 data type name

* fix: bitmap64 add type casting

* fix: bitmap64_range_int_array returns random array items

* fix: bitmap64_get_int_array returning a full range of bitmap

* chore: editor config

* refactor: unit tests

* fix: free test resources

+++ src/data-structure.h
@@ -70,9 +70,9 @@ Bitmap* bitmap_not(const Bitmap* bitmap);
 Bitmap* bitmap_from_int_array(size_t n, const uint32_t* array);
 Bitmap64* bitmap64_from_int_array(size_t n, const uint64_t* array);
 uint32_t* bitmap_get_int_array(const Bitmap* bitmap, size_t* n);
-uint64_t* bitmap64_get_int_array(const Bitmap64* bitmap, size_t* n);
+uint64_t* bitmap64_get_int_array(const Bitmap64* bitmap, uint64_t* n);
 uint32_t* bitmap_range_int_array(const Bitmap* bitmap, size_t offset, size_t n);
-uint64_t* bitmap64_range_int_array(const Bitmap64* bitmap, size_t offset, size_t n);
+uint64_t* bitmap64_range_int_array(const Bitmap64* bitmap, uint64_t offset, uint64_t n);
 void bitmap_free_int_array(uint32_t* array);
 void bitmap64_free_int_array(uint64_t* array);
 /**
