+++ src/redis-roaring.c
@@ -271,12 +271,12 @@ int R64GetIntArrayCommand(RedisModuleCtx* ctx, RedisModuleString** argv, int arg
 
   Bitmap64* bitmap = (type == REDISMODULE_KEYTYPE_EMPTY) ? NULL : RedisModule_ModuleTypeGetValue(key);
 
-  size_t n = 0;
+  uint64_t n = 0;
   uint64_t* array = bitmap == NULL ? NULL : bitmap64_get_int_array(bitmap, &n);
 
   RedisModule_ReplyWithArray(ctx, n);
 
-  for (size_t i = 0; i < n; i++) {
+  for (uint64_t i = 0; i < n; i++) {
     ReplyWithUint64(ctx, array[i]);
   }
 
@@ -316,16 +316,17 @@ int R64RangeIntArrayCommand(RedisModuleCtx* ctx, RedisModuleString** argv, int a
 
   Bitmap64* bitmap = (type == REDISMODULE_KEYTYPE_EMPTY) ? NULL : RedisModule_ModuleTypeGetValue(key);
   uint64_t* array = NULL;
-  size_t n = 0;
+  uint64_t n = 0;
   if (bitmap != NULL) {
     unsigned long long count = (unsigned long long) bitmap64_get_cardinality(bitmap);
+
     if (start > count - 1 || start > end) {
       n = 0;
     } else {
       n = end - start + 1;
     }
     if (n > 0) {
-      array = bitmap64_range_int_array(bitmap, (size_t) start, n);
+      array = bitmap64_range_int_array(bitmap, start, n);
     }
   } else {
     n = 0;
